refactored somewhat
