+++ mobile/src/main/AndroidManifest.xml
@@ -8,14 +8,14 @@
         android:label="@string/app_name"
         android:theme="@style/AppTheme">
 
-        <activity android:name=".MainActivity">
+        <activity android:name=".ui.main.MainActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
         <service
-            android:name=".PresentationService"
+            android:name=".ui.cast.PresentationService"
             android:exported="false" />
 
         <meta-data
