--- mobile/src/main/java/be/thmbc/castextendedscreen/CastEnabledActivity.java
@@ -1,9 +1,0 @@
-package be.thmbc.castextendedscreen;
-
-import android.support.v7.app.AppCompatActivity;
-
-/**
- * Created by maarten on 17/05/16.
- */
-public class CastEnabledActivity extends AppCompatActivity {
-}
